datahub-integrations-service/pyproject.toml
[build-system]
requires = ["setuptools>=78.1.1"]
build-backend = "setuptools.build_meta"

[project]
name = "acryl-datahub-integrations"
version = "1!0.0.0.dev0"
description = ""
authors = [
	{ name="Acryl Data", email="eng@acryl.io" },
]
requires-python = ">=3.11"

dependencies = [
	# 1.3.0: Pydantic v2 was introduced in 1.2.0.10 and later RCs fixed in OSS some missing code from /metadata-ingestion SaaS
	"acryl-datahub[base,sql-parser,snowflake-slim,bigquery-slim]>=1.3.1.5rc10",
	"acryl-datahub-actions[kafka,tag_propagation,term_propagation,doc_propagation,snowflake_tag_propagation]>=1.3.1.5rc10",

	"fastapi>=0.103.1",
	"uvicorn[standard]>=0.23.2",
	"pydantic>=2.10.0",
	"loguru",
	"sentry-sdk[loguru]",
	"build",
	# Slack.
	"slack-bolt",
	"inflect",
	# Teams Bot Framework.
	"botbuilder-core",
	"botbuilder-schema",
	"botframework-connector",
	# MCP
	"fastmcp",
	"markdown_to_mrkdwn",
	# Actions.
	"httpx",
	"anyio>=4",
	"asyncer",
	"reactpy",
	"reactpy-router",
	# AI.
	"tenacity",
	"boto3",
	"mlflow-skinny>=3.3.1,<4",
	"sagemaker-mlflow==0.2.0",
	"requests-auth-aws-sigv4==0.7",
	"cryptography>=43.0.0,<46",
	"pyOpenSSL>=25.1.0",
	"litellm>=1.0.0",
	"beautifulsoup4>=4.13.5",
	"json-repair",  # For parsing Bedrock's malformed JSON in structured output
	# LLM provider abstractions via langchain
	"langchain-core>=1.2.5",  # CVE-2025-68664: Serialization injection vulnerability
	"langchain>=1.0.5",
	"langchain-openai>=1.0.2",
	"langchain-google-vertexai>=3.0.2",
	# Email notifications.
	"sendgrid",
	# Reporting.
	"duckdb",
	"pandas",
	"boto3",
	"pyarrow",
	"jinja2",
	"cachetools",
	# dbt sync.
	"pygitops",
	"PyGithub",
	"GitPython",
	"h11>=0.16",
	"azure-identity==1.21.0",
	"aws-msk-iam-sasl-signer-python==1.0.2",
	"msal>=1.32.3",
	# Databricks sync -> conaction hangs forever on malformed http_path with newer databricks-sql-connector
	"databricks-sql-connector>=2.8.0,<3.0.0",
	# Debugging tools.
	"debugpy", # Enables remote debugging. Used in dev env
	"py-spy",
	"inotify",
	# Safe regex engine (prevents ReDoS attacks from user-supplied patterns)
	"google-re2",
	# OpenTelemetry for observability.
	"opentelemetry-api>=1.22.0",
	"opentelemetry-sdk>=1.22.0",
	"opentelemetry-instrumentation-fastapi>=0.43b0",
	"opentelemetry-instrumentation-httpx>=0.43b0",
	"opentelemetry-instrumentation-system-metrics>=0.43b0",
	"opentelemetry-exporter-prometheus>=0.43b0",
	"opentelemetry-exporter-otlp-proto-grpc>=1.22.0",
	"prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
	"acryl-datahub[testing-utils]",
	"setuptools",  # for pkg_resources
	"ruff==0.11.4",
	"pytest",
	"pytest-asyncio",
	"mypy",
	"pytest-cov",
	"asgi-lifespan",
	"httpx",
	"httpx-sse",
	"types-setuptools",
	"types-cachetools",
	"types-PyYAML",
	"types-requests>=2.28.11.6,<=2.31.0.3",
	"mypy-boto3-bedrock-runtime",
	"faker",
	"freezegun",
	"typer",
	"pandas-stubs",
	"ratelimit-stubs",
	"diskcache-stubs",
	"types-jmespath",
	"sqlalchemy-stubs",
	"types-psutil",
	# Iteration
	"papermill",
	"nbconvert",
	"json5",
	"streamlit",
	"streamlit-aggrid",
	"tiktoken",
	"mlflow",
	"diskcache",
	"seaborn",
	"matplotlib",
	"ipykernel",
]

[project.entry-points."datahub_actions.source.plugins"]
replay = "datahub_integrations.actions.replay_source.ReplayEventSource"

[project.entry-points."datahub_actions.action.plugins"]
record = "datahub_integrations.actions.recording_action.RecordingAction"
dbt_sync_back = "datahub_integrations.dbt.dbt_sync_back:DbtSyncBackAction"
ai_term_suggestion = "datahub_integrations.gen_ai.term_suggestion_action:TermSuggestionAction"
snowflake_metadata_sync = "datahub_integrations.propagation.snowflake.metadata_sync_action:SnowflakeMetadataSyncAction"

[tool.ruff]
target-version = "py311"
line-length = 88
exclude = [
	".git",
	"venv",
	"__pycache__",
	"experiments/*-private/",
	"experiments/*/archive/",
	"scripts/forms/*.py",
]

[tool.ruff.lint.isort]
section-order = ["future", "patch", "standard-library", "third-party", "first-party", "local-folder"]
sections = { "patch" = ["datahub.utilities._markupsafe_compat", "datahub.sql_parsing._sqlglot_patch", "datahub_integrations.gen_ai.mlflow_init", "datahub_integrations.experimentation.ai_init"] }
combine-as-imports = true

[tool.ruff.lint]
extend-select = [
	"ASYNC",
	"B",      # flake8-bugbear
	"E",      # pycodestyle errors
	"F",      # pyflakes
	"G010",   # logging.warn -> logging.warning
	"I",      # isort
	"INP001",
	"TID",    # flake8-tidy-imports
	"RUF100", # unused-noqa
]
extend-ignore = [
	'E501', # Ignore line length, since black handles that.
	'D203', # Ignore 1 blank line required before class docstring.
]
exclude = [
	"experiments/**/*.ipynb",
]

[tool.ruff.lint.extend-per-file-ignores]
"experiments/**.py" = ["INP001", "E402"]
"scripts/**.py" = ["B008"]

[tool.pyright]
extraPaths = ['tests']

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "asyncio: mark a test as an asyncio coroutine.",
    "integration: mark a test as an integration test requiring external dependencies."
]
norecursedirs = [
    "scripts/teams/install/datahub-router"
]
